BlockPoolMemoryManager
git-svn-id: https://g3d.svn.sourceforge.net/svnroot/g3d@257 ef0428c2-0c16-0410-9f3f-d710ca1f8061

+++ G3D9/G3D.lib/source/FileSystem.cpp
@@ -879,7 +879,7 @@ void FilePath::parse
 
 std::string FilePath::expandEnvironmentVariables(const std::string& path) {
     // Search for pattern
-    int end = path.find_first_of('$', 0);
+    size_t end = path.find_first_of('$', 0);
     if (end == std::string::npos) {
         // Pattern does not exist
         return path;
